Revert "maintain"
This reverts commit 96501429852013c8fb710796fde8d19414ccf4f7.

+++ framework/network/command/commandserver/lib/include/libNetFrame_CmdServer/Heart.h
@@ -19,7 +19,7 @@ typedef NETHANDLE		_HASH;
 		struct  stHeartInfo
 		{
 			CmdSession_sptr _Session_ptr;
-			unsigned long	_u32CurrentTimer;
+			UInt32	_u32CurrentTimer;
 			bool	_isValid;
 			
 			stHeartInfo( const CmdSession_sptr& Session_ptr )
@@ -101,7 +101,7 @@ typedef NETHANDLE		_HASH;
 			{
 				s_Lock.Lock();
 				stHeartInfo* pHeart = s_List.current();
-				if( (unsigned long)GetTickCount() - pHeart->_u32CurrentTimer > HEART_INTERVAL_TIME * HEART_RESEND_TIME + 5000 )
+				if( GetTickCount() - pHeart->_u32CurrentTimer > HEART_INTERVAL_TIME * HEART_RESEND_TIME + 5000 )
 				{
 					s_List.pop();
 					if( pHeart->_u32CurrentTimer != 0 )
